History score decay (#386)
Fix history scores overflowing. Formula from Ethereal.

Testing with [-5, 0]:

ELO | 11.05 +- 7.87 (95%)
SPRT | 10.0+0.1s Threads=1 Hash=32MB
LLR | 2.95 (-2.94, 2.94) [-5.00, 0.00]
Games | N: 3648 W: 947 L: 831 D: 1870
http://chess.grantnet.us/test/9366/

ELO | 10.39 +- 6.89 (95%)
SPRT | 60.0+0.6s Threads=1 Hash=128MB
LLR | 2.95 (-2.94, 2.94) [-5.00, 0.00]
Games | N: 3680 W: 750 L: 640 D: 2290
http://chess.grantnet.us/test/9367/

+++ src/threads.h
@@ -49,7 +49,7 @@ typedef struct Thread {
 
     Stack ss[128];
 
-    int history[2][64][64];
+    int16_t history[2][64][64];
 
     // Anything below here is not zeroed out between searches
     Position pos;
